Build the site as a test on PRs
This new action performs most of the same steps as the pages action, but
it does aim to publish anything. The goal here is to ensure that we have
some automated verification that new changes do not break the site
before merging.

This also configures the jinja environment to process templates more
strictly, causing erroneous template tags to raise an error in the build
process instead of failing silently.