Ensure that pre-commit is run for PRs and on the main branch
This change uses pre-commit as a linter / validation tool and will cause
PR checks to fail any time a pre-commit task raises an error.

I plan to follow this up with a GitHub action that applies any automated
formatting and fixes without requiring end-users to take additional
action on a pull request, but this validation action is a good first
step.