Change .rss extension to .xml
My browser (at least when serving static files generated locally, tested
with both Firefox and Chrome) does not handle the link ending with
`.rss` as a normal xml file and instead attempts to download it. This
change overrides the base `RSSFeed` class to change the file extension
to `.xml`, which is handled and displayed as I would expect an xml
document to be.